DPL Analysis: add RNTuple arrow::Dataset support
As part of the changes, move the actual logic which serialises /
deserialised things to plugins so that we do not need to depend
or RNTuple in production code.

Include an initial converter to go from AO2Ds to RNTuple based files.
