+++ Framework/Core/CMakeLists.txt
@@ -270,6 +270,10 @@ o2_add_test(Timers NAME test_Framework_test_Timers
             LABELS framework
             PUBLIC_LINK_LIBRARIES O2::Framework)
 
+o2_add_executable(framework-ao2d-to-ao3d
+            SOURCES test/o2AO2DToAO3D.cxx
+            PUBLIC_LINK_LIBRARIES O2::Framework)
+
 # FIXME: make this a proper test, when it actually does not hang.
 o2_add_executable(test-framework-ConsumeWhenAllOrdered
             SOURCES test/test_ConsumeWhenAllOrdered.cxx
@@ -299,6 +303,7 @@ add_executable(o2-test-framework-root
 target_link_libraries(o2-test-framework-root PRIVATE O2::Framework)
 target_link_libraries(o2-test-framework-root PRIVATE O2::Catch2)
 target_link_libraries(o2-test-framework-root PRIVATE ROOT::ROOTDataFrame)
+target_link_libraries(o2-test-framework-root PRIVATE ROOT::ROOTNTuple)
 set_property(TARGET o2-test-framework-root PROPERTY RUNTIME_OUTPUT_DIRECTORY ${outdir})
 add_test(NAME framework:root COMMAND o2-test-framework-root --skip-benchmarks)
 add_test(NAME framework:crash COMMAND  sh -e -c "PATH=${CMAKE_RUNTIME_OUTPUT_DIRECTORY}:$PATH ${CMAKE_CURRENT_LIST_DIR}/test/test_AllCrashTypes.sh")
