+++ Framework/Core/include/Framework/RootArrowFilesystem.h
@@ -11,6 +11,7 @@
 #ifndef O2_FRAMEWORK_ROOT_ARROW_FILESYSTEM_H_
 #define O2_FRAMEWORK_ROOT_ARROW_FILESYSTEM_H_
 
+#include <TBufferFile.h>
 #include <arrow/dataset/type_fwd.h>
 #include <arrow/dataset/file_base.h>
 #include <arrow/filesystem/type_fwd.h>
@@ -18,22 +19,11 @@
 #include <memory>
 
 class TFile;
-class TBranch;
-class TTree;
 class TBufferFile;
 class TDirectoryFile;
 
 namespace o2::framework
 {
-
-class TTreeFileWriteOptions : public arrow::dataset::FileWriteOptions
-{
- public:
-  TTreeFileWriteOptions(std::shared_ptr<arrow::dataset::FileFormat> format)
-    : FileWriteOptions(format)
-  {
-  }
-};
 
 // This is to avoid having to implement a bunch of unimplemented methods
 // for all the possible virtual filesystem we can invent on top of ROOT
@@ -79,46 +69,43 @@ class VirtualRootFileSystemBase : public arrow::fs::FileSystem
     const std::shared_ptr<const arrow::KeyValueMetadata>& metadata) override;
 };
 
-// A filesystem which allows me to get a TTree
-class TTreeFileSystem : public VirtualRootFileSystemBase
-{
- public:
-  ~TTreeFileSystem() override;
-
-  std::shared_ptr<VirtualRootFileSystemBase> GetSubFilesystem(arrow::dataset::FileSource source) override
-  {
-    return std::dynamic_pointer_cast<VirtualRootFileSystemBase>(shared_from_this());
-  };
-
-  arrow::Result<std::shared_ptr<arrow::io::OutputStream>> OpenOutputStream(
-    const std::string& path,
-    const std::shared_ptr<const arrow::KeyValueMetadata>& metadata) override;
-
-  virtual TTree* GetTree(arrow::dataset::FileSource source) = 0;
+struct RootArrowFactory final {
+  std::function<std::shared_ptr<arrow::dataset::FileWriteOptions>()> options = nullptr;
+  std::function<std::shared_ptr<arrow::dataset::FileFormat>()> format = nullptr;
+  std::function<std::shared_ptr<VirtualRootFileSystemBase>(void*)> getSubFilesystem = nullptr;
 };
 
-class SingleTreeFileSystem : public TTreeFileSystem
-{
- public:
-  SingleTreeFileSystem(TTree* tree)
-    : TTreeFileSystem(),
-      mTree(tree)
-  {
-  }
+struct RootArrowFactoryPlugin {
+  virtual RootArrowFactory* create() = 0;
+};
 
-  std::string type_name() const override
-  {
-    return "ttree";
-  }
+// A registry for all the possible ways of encoding a table in a TFile
+struct RootObjectReadingCapability {
+  // The unique name of this capability
+  std::string name = "unknown";
+  // Given a TFile, return the object which this capability support
+  // Use a void * in order not to expose the kind of object to the
+  // generic reading code. This is also where we load the plugin
+  // which will be used for the actual creation.
+  std::function<void*(TDirectoryFile* file, std::string const& path)> getHandle;
+  // Same as the above, but uses a TBufferFile as storage
+  std::function<void*(TBufferFile*, std::string const&)> getBufferHandle;
+  // This must be implemented to load the actual RootArrowFactory plugin which
+  // implements this capability. This way the detection of the file format
+  // (via get handle) does not need to know about the actual code which performs
+  // the serialization (and might depend on e.g. RNTuple).
+  std::function<RootArrowFactory&()> factory;
+};
 
-  TTree* GetTree(arrow::dataset::FileSource) override
-  {
-    // Simply return the only TTree we have
-    return mTree;
-  }
+struct RootObjectReadingCapabilityPlugin {
+  virtual RootObjectReadingCapability* create() = 0;
+};
 
- private:
-  TTree* mTree;
+// This acts as registry of all the capabilities (i.e. the ability to
+// associate a given object in a root file to the serialization plugin) and
+// the factory (i.e. the serialization plugin)
+struct RootObjectReadingFactory {
+  std::vector<RootObjectReadingCapability> capabilities;
 };
 
 class TFileFileSystem : public VirtualRootFileSystemBase
@@ -126,7 +113,7 @@ class TFileFileSystem : public VirtualRootFileSystemBase
  public:
   arrow::Result<arrow::fs::FileInfo> GetFileInfo(const std::string& path) override;
 
-  TFileFileSystem(TDirectoryFile* f, size_t readahead);
+  TFileFileSystem(TDirectoryFile* f, size_t readahead, RootObjectReadingFactory&);
 
   std::string type_name() const override
   {
@@ -147,12 +134,13 @@ class TFileFileSystem : public VirtualRootFileSystemBase
 
  private:
   TDirectoryFile* mFile;
+  RootObjectReadingFactory& mObjectFactory;
 };
 
 class TBufferFileFS : public VirtualRootFileSystemBase
 {
  public:
-  TBufferFileFS(TBufferFile* f);
+  TBufferFileFS(TBufferFile* f, RootObjectReadingFactory&);
 
   arrow::Result<arrow::fs::FileInfo> GetFileInfo(const std::string& path) override;
   std::string type_name() const override
@@ -165,68 +153,7 @@ class TBufferFileFS : public VirtualRootFileSystemBase
  private:
   TBufferFile* mBuffer;
   std::shared_ptr<VirtualRootFileSystemBase> mFilesystem;
-};
-
-class TTreeFileFragment : public arrow::dataset::FileFragment
-{
- public:
-  TTreeFileFragment(arrow::dataset::FileSource source,
-                    std::shared_ptr<arrow::dataset::FileFormat> format,
-                    arrow::compute::Expression partition_expression,
-                    std::shared_ptr<arrow::Schema> physical_schema)
-    : FileFragment(std::move(source), std::move(format), std::move(partition_expression), std::move(physical_schema))
-  {
-  }
-};
-
-class TTreeFileFormat : public arrow::dataset::FileFormat
-{
-  size_t& mTotCompressedSize;
-  size_t& mTotUncompressedSize;
-
- public:
-  TTreeFileFormat(size_t& totalCompressedSize, size_t& totalUncompressedSize)
-    : FileFormat({}),
-      mTotCompressedSize(totalCompressedSize),
-      mTotUncompressedSize(totalUncompressedSize)
-  {
-  }
-
-  ~TTreeFileFormat() override = default;
-
-  std::string type_name() const override
-  {
-    return "ttree";
-  }
-
-  bool Equals(const FileFormat& other) const override
-  {
-    return other.type_name() == this->type_name();
-  }
-
-  arrow::Result<bool> IsSupported(const arrow::dataset::FileSource& source) const override
-  {
-    auto fs = std::dynamic_pointer_cast<VirtualRootFileSystemBase>(source.filesystem());
-    auto subFs = fs->GetSubFilesystem(source);
-    if (std::dynamic_pointer_cast<TTreeFileSystem>(subFs)) {
-      return true;
-    }
-    return false;
-  }
-
-  arrow::Result<std::shared_ptr<arrow::Schema>> Inspect(const arrow::dataset::FileSource& source) const override;
-  /// \brief Create a FileFragment for a FileSource.
-  arrow::Result<std::shared_ptr<arrow::dataset::FileFragment>> MakeFragment(
-    arrow::dataset::FileSource source, arrow::compute::Expression partition_expression,
-    std::shared_ptr<arrow::Schema> physical_schema) override;
-
-  arrow::Result<std::shared_ptr<arrow::dataset::FileWriter>> MakeWriter(std::shared_ptr<arrow::io::OutputStream> destination, std::shared_ptr<arrow::Schema> schema, std::shared_ptr<arrow::dataset::FileWriteOptions> options, arrow::fs::FileLocator destination_locator) const override;
-
-  std::shared_ptr<arrow::dataset::FileWriteOptions> DefaultWriteOptions() override;
-
-  arrow::Result<arrow::RecordBatchGenerator> ScanBatchesAsync(
-    const std::shared_ptr<arrow::dataset::ScanOptions>& options,
-    const std::shared_ptr<arrow::dataset::FileFragment>& fragment) const override;
+  RootObjectReadingFactory& mObjectFactory;
 };
 
 // An arrow outputstream which allows to write to a TDirectoryFile.
@@ -253,33 +180,6 @@ class TDirectoryFileOutputStream : public arrow::io::OutputStream
 
  private:
   TDirectoryFile* mDirectory;
-};
-
-// An arrow outputstream which allows to write to a TTree. Eventually
-// with a prefix for the branches.
-class TTreeOutputStream : public arrow::io::OutputStream
-{
- public:
-  TTreeOutputStream(TTree*, std::string branchPrefix);
-
-  arrow::Status Close() override;
-
-  arrow::Result<int64_t> Tell() const override;
-
-  arrow::Status Write(const void* data, int64_t nbytes) override;
-
-  bool closed() const override;
-
-  TBranch* CreateBranch(char const* branchName, char const* sizeBranch);
-
-  TTree* GetTree()
-  {
-    return mTree;
-  }
-
- private:
-  TTree* mTree;
-  std::string mBranchPrefix;
 };
 
 } // namespace o2::framework
