+++ Framework/Core/test/test_Root2ArrowTable.cxx
@@ -14,6 +14,7 @@
 #include "Framework/TableBuilder.h"
 #include "Framework/RootTableBuilderHelpers.h"
 #include "Framework/ASoA.h"
+#include "Framework/PluginManager.h"
 #include "../src/ArrowDebugHelpers.h"
 
 #include <ROOT/RDataFrame.hxx>
@@ -26,6 +27,13 @@
 #include <TTree.h>
 #include <TRandom.h>
 #include <TFile.h>
+#include <ROOT/RField.hxx>
+#include <ROOT/RNTuple.hxx>
+#include <ROOT/RNTupleDescriptor.hxx>
+#include <ROOT/RNTupleModel.hxx>
+#include <ROOT/RNTupleReader.hxx>
+#include <ROOT/RNTupleUtil.hxx>
+#include <ROOT/RNTupleWriter.hxx>
 #include <memory>
 
 #include <arrow/array/array_primitive.h>
@@ -232,10 +240,31 @@ TEST_CASE("RootTree2Fragment")
   file->WriteObjectAny(&t1, t1.Class());
   auto* fileRead = new TBufferFile(TBuffer::kRead, file->BufferSize(), file->Buffer(), false, nullptr);
 
-  size_t totalSizeCompressed = 0;
-  size_t totalSizeUncompressed = 0;
-  auto format = std::make_shared<TTreeFileFormat>(totalSizeCompressed, totalSizeUncompressed);
-  auto fs = std::make_shared<TBufferFileFS>(fileRead);
+  std::vector<char const*> capabilitiesSpecs = {
+    "O2Framework:RNTupleObjectReadingCapability",
+    "O2Framework:TTreeObjectReadingCapability",
+  };
+
+  std::vector<LoadablePlugin> plugins;
+  for (auto spec : capabilitiesSpecs) {
+    auto morePlugins = PluginManager::parsePluginSpecString(spec);
+    for (auto& extra : morePlugins) {
+      plugins.push_back(extra);
+    }
+  }
+  REQUIRE(plugins.size() == 2);
+
+  RootObjectReadingFactory factory;
+  std::vector<char const*> configDiscoverySpec = {};
+  PluginManager::loadFromPlugin<RootObjectReadingCapability, RootObjectReadingCapabilityPlugin>(plugins, factory.capabilities);
+  REQUIRE(factory.capabilities.size() == 2);
+  REQUIRE(factory.capabilities[0].name == "rntuple");
+  REQUIRE(factory.capabilities[1].name == "ttree");
+
+  // Plugins are hardcoded for now...
+  auto format = factory.capabilities[1].factory().format();
+
+  auto fs = std::make_shared<TBufferFileFS>(fileRead, factory);
 
   arrow::dataset::FileSource source("p", fs);
   REQUIRE(format->IsSupported(source) == true);
@@ -439,10 +468,34 @@ TEST_CASE("RootTree2Dataset")
   }
   f->Write();
 
-  size_t totalSizeCompressed = 0;
-  size_t totalSizeUncompressed = 0;
-  auto format = std::make_shared<TTreeFileFormat>(totalSizeCompressed, totalSizeUncompressed);
-  auto fs = std::make_shared<TFileFileSystem>(f, 50 * 1024 * 1024);
+  std::vector<char const*> capabilitiesSpecs = {
+    "O2Framework:RNTupleObjectReadingCapability",
+    "O2Framework:TTreeObjectReadingCapability",
+  };
+
+  RootObjectReadingFactory factory;
+
+  std::vector<LoadablePlugin> plugins;
+  for (auto spec : capabilitiesSpecs) {
+    auto morePlugins = PluginManager::parsePluginSpecString(spec);
+    for (auto& extra : morePlugins) {
+      plugins.push_back(extra);
+    }
+  }
+  REQUIRE(plugins.size() == 2);
+
+  PluginManager::loadFromPlugin<RootObjectReadingCapability, RootObjectReadingCapabilityPlugin>(plugins, factory.capabilities);
+
+  REQUIRE(factory.capabilities.size() == 2);
+  REQUIRE(factory.capabilities[0].name == "rntuple");
+  REQUIRE(factory.capabilities[1].name == "ttree");
+
+  // Plugins are hardcoded for now...
+  auto rNtupleFormat = factory.capabilities[0].factory().format();
+  auto format = factory.capabilities[1].factory().format();
+
+  auto fs = std::make_shared<TFileFileSystem>(f, 50 * 1024 * 1024, factory);
+
   arrow::dataset::FileSource source("DF_2/tracks", fs);
   REQUIRE(format->IsSupported(source) == true);
   auto schemaOpt = format->Inspect(source);
@@ -464,7 +517,7 @@ TEST_CASE("RootTree2Dataset")
   validateContents(*result);
 
   auto* output = new TMemFile("foo", "RECREATE");
-  auto outFs = std::make_shared<TFileFileSystem>(output, 0);
+  auto outFs = std::make_shared<TFileFileSystem>(output, 0, factory);
 
   // Open a stream at toplevel
   auto destination = outFs->OpenOutputStream("/", {});
@@ -503,4 +556,38 @@ TEST_CASE("RootTree2Dataset")
     REQUIRE((*resultWritten)->num_rows() == 100);
     validateContents(*resultWritten);
   }
+  arrow::fs::FileLocator rnTupleLocator{outFs, "/rntuple"};
+  // We write an RNTuple in the same TMemFile, using /rntuple as a location
+  auto rntupleDestination = std::dynamic_pointer_cast<TDirectoryFileOutputStream>(*destination);
+
+  {
+    auto rNtupleWriter = rNtupleFormat->MakeWriter(*destination, schema, {}, rnTupleLocator);
+    auto rNtupleSuccess = rNtupleWriter->get()->Write(*result);
+    REQUIRE(rNtupleSuccess.ok());
+  }
+
+  // And now we can read back the RNTuple into a RecordBatch
+  arrow::dataset::FileSource writtenRntupleSource("/rntuple", outFs);
+  auto newRNTupleFS = outFs->GetSubFilesystem(writtenRntupleSource);
+
+  REQUIRE(rNtupleFormat->IsSupported(writtenRntupleSource) == true);
+
+  auto rntupleSchemaOpt = rNtupleFormat->Inspect(writtenRntupleSource);
+  REQUIRE(rntupleSchemaOpt.ok());
+  auto rntupleSchemaWritten = *rntupleSchemaOpt;
+  REQUIRE(validateSchema(rntupleSchemaWritten));
+
+  auto rntupleFragmentWritten = rNtupleFormat->MakeFragment(writtenRntupleSource, {}, rntupleSchemaWritten);
+  REQUIRE(rntupleFragmentWritten.ok());
+  auto rntupleOptionsWritten = std::make_shared<arrow::dataset::ScanOptions>();
+  rntupleOptionsWritten->dataset_schema = rntupleSchemaWritten;
+  auto rntupleScannerWritten = rNtupleFormat->ScanBatchesAsync(rntupleOptionsWritten, *rntupleFragmentWritten);
+  REQUIRE(rntupleScannerWritten.ok());
+  auto rntupleBatchesWritten = (*rntupleScannerWritten)();
+  auto rntupleResultWritten = rntupleBatchesWritten.result();
+  REQUIRE(rntupleResultWritten.ok());
+  REQUIRE((*rntupleResultWritten)->columns().size() == 10);
+  REQUIRE(validateSchema((*rntupleResultWritten)->schema()));
+  REQUIRE((*rntupleResultWritten)->num_rows() == 100);
+  REQUIRE(validateContents(*rntupleResultWritten));
 }
